Implement a React collapse component to override NU Global Marketing collapse jQuery plugin
