+++ package.json
@@ -18,7 +18,9 @@
     "node-sass-chokidar": "^1.1.0",
     "npm-run-all": "^4.1.2",
     "prop-types": "^15.6.1",
+    "rc-collapse": "^1.10.0",
     "react": "^16.0.0",
+    "react-collapsible": "^2.3.0",
     "react-copy-to-clipboard": "^5.0.1",
     "react-dom": "^16.0.0",
     "react-redux": "^5.0.7",
