+++ src/components/ItemDetail/ItemDetailCarousels.js
@@ -1,16 +1,12 @@
 import React from 'react';
 import CarouselSection from '../CarouselSection';
 import PropTypes from 'prop-types';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import ThisItem from './ThisItem';
+import Collapsible from 'react-collapsible';
+import CollapsibleHeader from '../CollapsibleHeader';
 
 const ItemDetailCarousels = props => {
   const { adminSetItems, collectionItems, item } = props;
-  const styles = {
-    caret: {
-      fontSize: '6rem',
-      color: '#f0f0f0'
-    }
-  };
 
   const renderCollectionsCarousel = () => {
     const shouldDisplay =
@@ -21,46 +17,46 @@ const ItemDetailCarousels = props => {
     }
 
     return (
-      <div>
-        <h3 className="open">Collection</h3>
-        <CarouselSection
-          sectionTitle={item.collection[0].title[0]}
-          linkTo=""
-          items={collectionItems.items}
-          slidesPerView={6}
-          loading=""
-          error=""
-        />
-      </div>
+      <CarouselSection
+        sectionTitle={item.collection[0].title[0]}
+        linkTo=""
+        items={collectionItems.items}
+        slidesPerView={6}
+        loading=""
+        error=""
+      />
     );
   };
 
   return (
     <section className="contain-1120 item-section item-categories-wrapper">
       <h3>Library Division and Collections with this Item:</h3>
+      <div className="expander expander1">
+        <Collapsible
+          trigger={<CollapsibleHeader label="Library Division" />}
+          open={true}
+        >
+          {item &&
+            adminSetItems && (
+              <CarouselSection
+                sectionTitle={item.admin_set.title[0]}
+                linkTo=""
+                items={adminSetItems.items}
+                slidesPerView={6}
+                loading=""
+                error=""
+              />
+            )}
+        </Collapsible>
+        <Collapsible
+          trigger={<CollapsibleHeader label="Collection" />}
+          open={true}
+        >
+          {renderCollectionsCarousel()}
+        </Collapsible>
+      </div>
 
-      <div className="expander expander1" data-collapse="data-collapse">
-        {item && adminSetItems && <h3 className="open">Library Division</h3>}
-        {item &&
-          adminSetItems && (
-            <CarouselSection
-              sectionTitle={item.admin_set.title[0]}
-              linkTo=""
-              items={adminSetItems.items}
-              slidesPerView={6}
-              loading=""
-              error=""
-            />
-          )}
-        {renderCollectionsCarousel()}
-      </div>
-      <div className="this-item-wrapper">
-        <div>
-          <FontAwesomeIcon icon="caret-down" style={styles.caret} />
-        </div>
-        <p>This item</p>
-        <img src={item && item.thumbnail_url} alt={item.label} />
-      </div>
+      <ThisItem item={item} />
     </section>
   );
 };
