+++ src/components/ItemDetail/ThisItem.js
@@ -1,0 +1,29 @@
+import React from 'react';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import PropTypes from 'prop-types';
+
+const ThisItem = props => {
+  const { item } = props;
+  const styles = {
+    caret: {
+      fontSize: '6rem',
+      color: '#f0f0f0'
+    }
+  };
+
+  return (
+    <div className="this-item-wrapper">
+      <div>
+        <FontAwesomeIcon icon="caret-down" style={styles.caret} />
+      </div>
+      <p>This item</p>
+      <img src={item && item.thumbnail_url} alt={item && item.label} />
+    </div>
+  );
+};
+
+ThisItem.propTypes = {
+  item: PropTypes.object
+};
+
+export default ThisItem;
