+++ src/styles/sass/styles.scss
@@ -74,3 +74,4 @@
 @import "nulib-collections/sets";
 @import "nulib-collections/home";
 @import "nulib-collections/item-view";
+@import "nulib-collections/collapsible";
